backend/requirements.txt
asgiref             3.8.1
cffi                1.17.1
charset-normalizer  3.4.1
cryptography        44.0.0
Django              4.2.19
django-cors-headers 4.6.0
djangorestframework 3.15.2
dnspython           2.7.0
pdfminer.six        20231228
pdfplumber          0.11.5
pillow              11.1.0
pip                 25.0
pycparser           2.22
pymongo             4.11
pypdfium2           4.30.1
setuptools          58.1.0
sqlparse            0.5.3
typing_extensions   4.12.2
tzdata              2025.1